+++ README.md
@@ -107,7 +107,7 @@ I vow to:
 └── README.md
 ```
 
-The first stone is [`01_FOUNDATIONS/falsifiability.md`](./01_FOUNDATIONS/falsifiability.md) — criteria any claim to cognition must survive. Naming discipline is outlined in [`02_DESIGN_PRINCIPLES/no_god_words.md`](./02_DESIGN_PRINCIPLES/no_god_words.md). Architectural sketches live in [`03_CORE_ARCHITECTURE/distributed_agency.md`](./03_CORE_ARCHITECTURE/distributed_agency.md), and the core recursive loop is implemented in [`03_CORE_ARCHITECTURE/r2d2_core.py`](./03_CORE_ARCHITECTURE/r2d2_core.py) with its conceptual outline in [`03_CORE_ARCHITECTURE/r2d2_core.md`](./03_CORE_ARCHITECTURE/r2d2_core.md).
+The first stone is [`01_FOUNDATIONS/falsifiability.md`](./01_FOUNDATIONS/falsifiability.md) — criteria any claim to cognition must survive. Naming discipline is outlined in [`02_DESIGN_PRINCIPLES/no_god_words.md`](./02_DESIGN_PRINCIPLES/no_god_words.md). Architectural sketches live in [`03_CORE_ARCHITECTURE/distributed_agency.md`](./03_CORE_ARCHITECTURE/distributed_agency.md), and the core recursive loop is implemented in [`src/recursive_agency/r2d2_core.py`](./src/recursive_agency/r2d2_core.py) with its conceptual outline in [`03_CORE_ARCHITECTURE/r2d2_core.md`](./03_CORE_ARCHITECTURE/r2d2_core.md).
 
 ## Licensing
