+++ src/recursive_agency/__init__.py
@@ -1,0 +1,6 @@
+"""Recursive Agency package."""
+
+from .agency_engine import AgencyEngine
+from .r2d2_core import Capsule, R2D2Solver
+
+__all__ = ["AgencyEngine", "Capsule", "R2D2Solver"]
