+++ src/recursive_agency/r2d2_core.py
